refactor: invoice configuration fixed
