+++ src/components/invoice-configuration-form/InvoiceConfigurationForm.tsx
@@ -1,12 +1,17 @@
 import { Dialog } from '@/components/dialog/Dialog';
 import { useAppForm } from '@/hooks/use-app-form/useAppForm';
 import { cn } from '@/utils/classNamesUtils';
+import z from 'zod';
 import type { InvoiceConfigurationPersistSchema } from './invoiceConfigurationFormSchemas';
 import type { InvoiceConfigurationFormSchema } from './invoiceConfigurationFormSchemas';
 import { invoiceConfigurationFormSchema } from './invoiceConfigurationFormSchemas';
 import { InvoiceConfigurationFormFields } from './InvoiceConfigurationFormFields';
 
-export type ContractInvoiceConfigurationFormProps = {
+const wrappedInvoiceConfigurationFormSchema = z.object({
+	invoiceConfiguration: invoiceConfigurationFormSchema,
+});
+
+export type InvoiceConfigurationFormProps = {
 	defaultValues: InvoiceConfigurationFormSchema;
 	previewCompanyName: string;
 	previewReferenceDate: Date;
@@ -14,20 +19,20 @@ export type ContractInvoiceConfigurationFormProps = {
 	submitLabel: string;
 };
 
-export const ContractInvoiceConfigurationForm = ({
+export const InvoiceConfigurationForm = ({
 	defaultValues,
 	previewCompanyName,
 	previewReferenceDate,
 	onSuccess,
 	submitLabel,
-}: ContractInvoiceConfigurationFormProps) => {
+}: InvoiceConfigurationFormProps) => {
 	const form = useAppForm({
-		defaultValues,
+		defaultValues: { invoiceConfiguration: defaultValues },
 		validators: {
-			onChange: invoiceConfigurationFormSchema,
+			onChange: wrappedInvoiceConfigurationFormSchema,
 		},
 		onSubmit: ({ value }) => {
-			const { invoiceNumberingMode, ...rest } = value;
+			const { invoiceNumberingMode, ...rest } = value.invoiceConfiguration;
 			onSuccess({
 				...rest,
 				lastInvoiceNumber:
@@ -39,12 +44,13 @@ export const ContractInvoiceConfigurationForm = ({
 	return (
 		<form.Root
 			form={form}
-			schema={invoiceConfigurationFormSchema}
+			schema={wrappedInvoiceConfigurationFormSchema}
 			isLoading={false}
 		>
 			<Dialog.Body className={cn('gap-6')}>
 				<InvoiceConfigurationFormFields
 					form={form}
+					fields='invoiceConfiguration'
 					previewCompanyName={previewCompanyName}
 					previewReferenceDate={previewReferenceDate}
 				/>
